requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pdfplumber
langchain
langchain-google-genai
supabase
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
